Ipx updated code

+++ src/ipm/ipx/src/splitted_normal_matrix.cc
@@ -47,7 +47,7 @@ void SplittedNormalMatrix::Prepare(const Basis& basis, const double* colscale) {
     PermuteRows(N_, rowperm_inv_);
 
     // Scale columns of N.
-    for (Int k = 0; k < (Int) nonbasic_vars.size(); k++) {
+    for (Int k = 0; k < nonbasic_vars.size(); k++) {
         Int j = nonbasic_vars[k];
         double d = colscale[j];
         assert(std::isfinite(d));
